+++ src/main.rs
@@ -14,15 +14,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.*/
 //!
 //! ```rpick``` helps pick items from a list of choices, using various algorithms.
 
-use std::io;
-
 use structopt::StructOpt;
-use term::Terminal;
+
+mod cli;
 
 const CONFIG_FILE: &str = "rpick.yml";
 
 #[derive(StructOpt)]
-struct Cli {
+struct CliArgs {
     /// The category you wish to pick from.
     category: String,
     #[structopt(short, long, env = "RPICK_CONFIG")]
@@ -34,25 +33,15 @@ struct Cli {
 }
 
 fn main() {
-    let args = Cli::from_args();
+    let args = CliArgs::from_args();
     let config_path = get_config_file_path(&args);
     let config = rpick::read_config(&config_path);
     match config {
         Ok(config) => {
             let mut config = config;
-            let stdio = io::stdin();
-            let input = stdio.lock();
-            let mut output = io::stdout();
-            let color = match term::terminfo::TerminfoTerminal::new(&mut output) {
-                Some(term) => term.supports_color(),
-                None => false,
-            };
+            let ui = cli::CLI::new(args.verbose);
 
-            let mut engine = rpick::Engine::new(input, output, rand::thread_rng());
-            engine.color = color;
-            if args.verbose {
-                engine.verbose = true;
-            }
+            let mut engine = rpick::Engine::new(&ui);
             match engine.pick(&mut config, args.category) {
                 Ok(_) => match rpick::write_config(&config_path, config) {
                     Ok(_) => {}
@@ -78,7 +67,7 @@ fn main() {
 ///
 /// If the config flag is set in the given CLI args, that path is used. Otherwise, the default
 /// config name (CONFIG_FILE) is appended to the user's home config directory to form the path.
-fn get_config_file_path(args: &Cli) -> String {
+fn get_config_file_path(args: &CliArgs) -> String {
     match &args.config {
         Some(config) => config.clone(),
         None => {